rename client module to runner
